chore: file_manager_utils
