+++ engine/test/components/CMakeLists.txt
@@ -13,6 +13,7 @@ add_executable(${PROJECT_NAME}
   ${CMAKE_CURRENT_SOURCE_DIR}/../../services/download_service.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/../../database/models.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/../../utils/config_yaml_utils.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/../../utils/file_manager_utils.cc
 )
 
 find_package(Drogon CONFIG REQUIRED)
